requirements.txt
absl-py==2.1.0
addict==2.4.0
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
bokeh==3.1.1
Brotli==1.0.9
cachetools==5.3.3
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
cloudpickle==3.0.0
cmake==3.26.4
colorcet==3.1.0
comm==0.2.2
contourpy==1.1.0
crcmod==1.7
cryptography==42.0.5
cycler==0.11.0
dask==2023.5.0
datashader==0.15.2
datashape==0.5.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
docstring-parser==0.14.1
easydict==1.10
einops==0.7.0
essmc2==0.1.4
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
faiss-gpu==1.7.2
fastjsonschema==2.19.1
filelock==3.12.2
fonttools==4.40.0
fqdn==1.5.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
google-auth==2.31.0
google-auth-oauthlib==1.0.0
grpcio==1.64.1
h11==0.14.0
holoviews==1.17.1
httpcore==1.0.5
httpx==0.27.0
idna==3.4
imageio==2.34.2
importlib_metadata==7.1.0
importlib-resources==5.12.0
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==8.6.1
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.4
kornia==0.6.12
lazy_loader==0.4
linkify-it-py==2.0.3
lit==16.0.6
llvmlite==0.41.1
locket==1.0.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
multipledispatch==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook==7.1.3
notebook_shim==0.2.4
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-transforms==0.0.6
openpyxl==3.1.5
oss2==2.18.4
overrides==7.7.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
panel==1.2.3
param==2.1.1
parso==0.8.4
partd==1.4.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.20.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.12.0
pycparser==2.21
pycryptodome==3.20.0
pyct==0.5.0
Pygments==2.17.2
pynndescent==0.5.13
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
pyviz_comms==3.0.2
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.0.2
qtconsole==5.5.1
QtPy==2.4.1
randaugment==1.0.2
referencing==0.35.0
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.5.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tbparse==0.0.8
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboard-logger==0.1.0
tensorboardX==2.6.2.2
terminado==0.18.1
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.3.0
tomli==2.0.1
toolz==0.12.1
torch==2.0.1
torchaudio==2.0.2
torchnet==0.0.4
torchvision==0.15.2
tornado==6.4
tqdm==4.65.0
traitlets==5.14.3
triton==2.0.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.6.3
tzdata==2024.1
uc-micro-py==1.0.3
umap-learn==0.5.6
uri-template==1.3.0
urllib3==2.0.3
visdom==0.2.4
wandb==0.17.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.3
widgetsnbextension==4.0.10
xarray==2023.1.0
xyzservices==2024.6.0
yacs==0.1.8
yapf==0.32.0
yarg==0.1.9
zipp==3.15.0
